fixed tests  - moved some services  - added some providers  - altered others
