+++ frontend/build.gradle
@@ -1,3 +1,4 @@
+/*
 plugins {
     id "com.moowork.node" version "1.0.1"
 }
@@ -16,7 +17,9 @@ task assemble(dependsOn: ngBuild)
 
 task build(dependsOn: assemble)
 
+
 task ngTest(type: YarnTask, dependsOn: yarn_install) {
     args = ['run', 'ng', 'test', '--', '--watch=false', '--single-run', '--browser=PhantomJS']
 }
 task test(dependsOn: ngTest)
+*/
