+++ frontend/src/app/core/interfaces/auth.service.ts
