+++ frontend/src/app/core/mock/http/MockHttpInterceptor.ts
@@ -1,5 +1,5 @@
 import {Injectable} from '@angular/core';
-import {HttpInterceptor} from '../../../shared/http/HttpInterceptor';
+import {HttpInterceptor} from '../../http/HttpInterceptor';
 
 @Injectable()
 export class MockHttpInterceptor extends HttpInterceptor {
