+++ frontend/src/app/core/mock/ui/mock-confirm-dialog.service.ts
@@ -1,5 +1,5 @@
 import {Injectable} from '@angular/core';
-import {ConfirmDialogService} from '../../../shared/services/confirm-dialog.service';
+import {ConfirmDialogService} from '../../services/confirm-dialog.service';
 
 @Injectable()
 export class MockConfirmDialogService extends ConfirmDialogService {
