+++ frontend/src/app/core/mock/ui/mock-responsive-helper.service.ts
@@ -1,5 +1,5 @@
 import {Injectable} from '@angular/core';
-import {ResponsiveHelperService} from '../../../shared/services/ui/responsive-helper.service';
+import {ResponsiveHelperService} from '../../services/ui/responsive-helper.service';
 
 @Injectable()
 export class MockResponsiveHelperService extends ResponsiveHelperService {
