+++ frontend/src/app/oc-ui/oc-ui.module.ts
@@ -11,17 +11,17 @@ import {
     OCCollapsibleHeaderComponent
 } from './components/oc-collapsible/oc-collapsible';
 import {OCDividerDirective} from './directives/oc-divider.directive';
-import {MaterialModule} from '@angular/material';
 import {TranslateModule} from '@ngx-translate/core';
 import {OCNavComponent} from './components/oc-nav/oc-nav.component';
 import {RouterModule} from '@angular/router';
 import {FlexLayoutModule} from '@angular/flex-layout';
+import {OCMaterialModule} from '../oc-material.module';
 
 @NgModule({
     imports: [
         CommonModule,
         RouterModule,
-        MaterialModule,
+        OCMaterialModule,
         FlexLayoutModule,
         TranslateModule
     ],
